+++ src/defs/exceptions.py
