+++ src/seqvar_pvs1.py
@@ -7,7 +7,6 @@ import typer
 
 from src.api.annonars import AnnonarsClient
 from src.api.mehari import MehariClient
-from src.core.exceptions import InvalidAPIResposeError
 from src.defs.autopvs1 import (
     PVS1Prediction,
     PVS1PredictionSeqVarPath,
@@ -15,6 +14,7 @@ from src.defs.autopvs1 import (
     SeqvarConsequenceMapping,
     TranscriptInfo,
 )
+from src.defs.exceptions import InvalidAPIResposeError
 from src.defs.mehari import CdsPos, Exon, TranscriptGene, TranscriptSeqvar
 from src.defs.seqvar import SeqVar
